Changed the way of setting env variables.
The setting of environment variables was made a task of its own. What
happened is that a process was spawned by ruby and executed the EXPORT
bash command. However, this only changed the environment of the child
process and not the parent. Any further executed commands would still
only run with the old and previous environment.